refactor: Move hyprland-specific functions out of lib
Will update hyprland.rs to make it compile later.
