+++ src/lib.rs
@@ -1,10 +1,5 @@
 use clap::Parser;
 use cli::Cli;
-use hyprland::data::Client;
-use hyprland::data::Monitors;
-use hyprland::data::Workspace;
-use hyprland::data::Workspaces;
-use hyprland::prelude::*;
 
 mod cli;
 
@@ -28,21 +23,9 @@ pub struct WorkspaceState {
 /// Implementation for WorkspaceState
 impl WorkspaceState {
     /// Creates a new WorkspaceState
-    pub fn new() -> Self {
-        let workspaces = Workspaces::get().unwrap();
-        let mut monitor_ids: Vec<i32> = workspaces
-            .clone()
-            .filter(|x| x.monitor == Monitors::get().unwrap().find(|x| x.focused).unwrap().name)
-            .map(|x| x.id)
-            .filter(|x| x > &0)
-            .collect();
-        let mut occupied_ids: Vec<i32> = workspaces.map(|x| x.id).filter(|x| x > &0).collect();
-
-        monitor_ids.sort();
-        occupied_ids.sort();
-
+    pub fn new(current_id: i32, monitor_ids: Vec<i32>, occupied_ids: Vec<i32>) -> Self {
         Self {
-            current_id: Workspace::get_active().unwrap().id,
+            current_id,
             monitor_ids,
             occupied_ids,
         }
@@ -136,19 +119,4 @@ pub fn get_id() -> i32 {
     } else {
         get_next_id(state, no_empty || no_empty_after)
     }
-}
-
-/// Gets whether the current workspace is a special workspace or not.
-///
-/// This function works by getting which workspace the active window is in.
-///
-/// The if statement is used to make sure this function works when no window
-/// is the active window.
-pub fn is_special() -> bool {
-    if let Some(client) = Client::get_active().unwrap() {
-        let Client { workspace, .. } = client;
-        return workspace.name.contains("special");
-    }
-
-    false
 }
